+++ cmd/frontend/auth/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "non_public.go",
         "redirect.go",
         "reset_password.go",
-        "sign_out_cookie.go",
         "user.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/auth",
@@ -17,7 +16,6 @@ go_library(
         "//cmd/frontend/backend",
         "//cmd/frontend/internal/app/router",
         "//cmd/frontend/internal/app/ui/router",
-        "//cmd/frontend/internal/auth/session",
         "//cmd/frontend/internal/auth/userpasswd",
         "//internal/actor",
         "//internal/auth",
