+++ cmd/frontend/internal/app/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app",
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
-        "//cmd/frontend/auth",
         "//cmd/frontend/backend",
         "//cmd/frontend/internal/app/assetsutil",
         "//cmd/frontend/internal/app/debugproxies",
